+++ frb_example/pure_dart/lib/src/rust/api/pseudo_manual/method_twin_rust_async_sse.dart
@@ -26,7 +26,8 @@ class StaticGetterOnlyTwinRustAsyncSse extends RustOpaque {
   StaticGetterOnlyTwinRustAsyncSse.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  StaticGetterOnlyTwinRustAsyncSse.sseDecode(int ptr, int externalSizeOnNative)
+  StaticGetterOnlyTwinRustAsyncSse.sseDecode(
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
