+++ frb_example/pure_dart/lib/src/rust/api/pseudo_manual/rust_auto_opaque_twin_rust_async.dart
@@ -264,7 +264,7 @@ class BoxHelloTraitTwinRustAsync extends RustOpaque {
   BoxHelloTraitTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  BoxHelloTraitTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  BoxHelloTraitTwinRustAsync.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -283,7 +283,7 @@ class BoxMyTraitTwinRustAsync extends RustOpaque {
   BoxMyTraitTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  BoxMyTraitTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  BoxMyTraitTwinRustAsync.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -303,7 +303,7 @@ class EnumWithGoodAndOpaqueWithoutOptionTwinRustAsync extends RustOpaque {
       : super.dcoDecode(wire, _kStaticData);
 
   EnumWithGoodAndOpaqueWithoutOptionTwinRustAsync.sseDecode(
-      int ptr, int externalSizeOnNative)
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -322,7 +322,8 @@ class NonCloneSimpleEnumTwinRustAsync extends RustOpaque {
   NonCloneSimpleEnumTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  NonCloneSimpleEnumTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  NonCloneSimpleEnumTwinRustAsync.sseDecode(
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -341,7 +342,7 @@ class NonCloneSimpleTwinRustAsync extends RustOpaque {
   NonCloneSimpleTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  NonCloneSimpleTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  NonCloneSimpleTwinRustAsync.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -430,7 +431,7 @@ class OpaqueOneTwinRustAsync extends RustOpaque {
   OpaqueOneTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  OpaqueOneTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  OpaqueOneTwinRustAsync.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -449,7 +450,7 @@ class OpaqueTwoTwinRustAsync extends RustOpaque {
   OpaqueTwoTwinRustAsync.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  OpaqueTwoTwinRustAsync.sseDecode(int ptr, int externalSizeOnNative)
+  OpaqueTwoTwinRustAsync.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -471,7 +472,7 @@ class StructWithGoodAndOpaqueFieldWithoutOptionTwinRustAsync
       : super.dcoDecode(wire, _kStaticData);
 
   StructWithGoodAndOpaqueFieldWithoutOptionTwinRustAsync.sseDecode(
-      int ptr, int externalSizeOnNative)
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
